Incorporate review comment

+++ src/library_wasmfs.js
@@ -87,7 +87,7 @@ mergeInto(LibraryManager.library, {
       // Copy the file into a JS buffer on the heap.
       var buf = __wasmfs_read_file(pathName);
       // The integer length resides in the first 8 bytes of the buffer.
-      var length = {{{ makeGetValue('buf', '0', 'i32') }}};
+      var length = {{{ makeGetValue('buf', '0', 'i64') }}};
 
       // Default return type is binary.
       // The buffer contents exist 8 bytes after the returned pointer.
